Split company data file into single-entity files
Applies the same treatment as licence.data.js to company.data.js: split into company.data.js, address.data.js, and company-address.data.js, each returning a bare entity, composed by a new private _company() helper in licence.scenario.js. Also fixes a pluralization bug in the load fixture's _asArrays helper — the naive "append s unless it already ends in s" rule silently mispluralized company (companys) and left address/companyAddress unpluralized entirely, since both already end in s. licence-holder.spec.js updated to match.
